[ci] Traffic-Based App Rewards: consolidate integration tests
