--- apps/app/src/test/scala/org/lfdecentralizedtrust/splice/integration/tests/ScanAppRewardsComputationTimeBasedIntegrationTest.scala
@@ -1,105 +1,0 @@
-package org.lfdecentralizedtrust.splice.integration.tests
-
-import com.digitalasset.canton.HasExecutionContext
-import org.lfdecentralizedtrust.splice.config.ConfigTransforms.{
-  ConfigurableApp,
-  updateAutomationConfig,
-}
-import org.lfdecentralizedtrust.splice.integration.EnvironmentDefinition
-import org.lfdecentralizedtrust.splice.integration.tests.SpliceTests.IntegrationTestWithIsolatedEnvironment
-import org.lfdecentralizedtrust.splice.scan.automation.{
-  RewardComputationTrigger,
-  ScanAggregationTrigger,
-}
-import org.lfdecentralizedtrust.splice.util.*
-
-class ScanAppRewardsComputationTimeBasedIntegrationTest
-    extends IntegrationTestWithIsolatedEnvironment
-    with WalletTestUtil
-    with TimeTestUtil
-    with HasExecutionContext {
-
-  override def environmentDefinition: SpliceEnvironmentDefinition =
-    EnvironmentDefinition
-      .simpleTopology1SvWithSimTime(this.getClass.getSimpleName)
-      .withoutAutomaticRewardsCollectionAndAmuletMerging
-      .addConfigTransforms((_, config) =>
-        updateAutomationConfig(ConfigurableApp.Scan)(
-          _.withPausedTrigger[ScanAggregationTrigger]
-            .withPausedTrigger[RewardComputationTrigger]
-        )(config)
-      )
-      .withAdditionalSetup { implicit _env =>
-        advanceTimeForRewardAutomationToRunForCurrentRound
-      }
-
-  "reward accounting endpoints" should {
-    "return 404 when no data is available" in { implicit _env =>
-      clue("earliest-available returns None") {
-        sv1ScanBackend.getRewardAccountingEarliestAvailableRound() shouldBe None
-      }
-      clue("activity-totals returns None for round 0") {
-        sv1ScanBackend.getRewardAccountingActivityTotals(0L) shouldBe None
-      }
-    }
-
-    "return activity totals after aggregation trigger runs" in { implicit _env =>
-      // TODO(#4384): Update the time management here
-      //
-      // Advance enough ticks so that at least three rounds are fully closed and aggregated.
-      // (advanceRoundsToNextRoundOpening advances by 1 tick)
-      for (_ <- 1 to 7) {
-        advanceRoundsToNextRoundOpening
-        sv1ScanBackend.automation
-          .trigger[ScanAggregationTrigger]
-          .runOnce()
-          .futureValue
-      }
-
-      clue("Run the reward computation trigger") {
-        sv1ScanBackend.automation
-          .trigger[RewardComputationTrigger]
-          .runOnce()
-          .futureValue
-      }
-
-      val earliest = clue("Verify earliest available round is returned") {
-        val e = sv1ScanBackend.getRewardAccountingEarliestAvailableRound()
-        e.value shouldBe 1
-        e.value
-      }
-
-      clue("Verify activity totals for the aggregated round") {
-        val totals = sv1ScanBackend.getRewardAccountingActivityTotals(earliest)
-        totals.value.roundNumber shouldBe earliest
-        totals.value.activityRecordsCount should be > 0L
-      }
-
-      clue("Verify 404 for non-existent round") {
-        sv1ScanBackend.getRewardAccountingActivityTotals(earliest + 1) shouldBe None
-      }
-
-      clue("Verify root hash is available for the computed round") {
-        val rootHash = sv1ScanBackend.getRewardAccountingRootHash(earliest)
-        rootHash shouldBe defined
-        rootHash.value.roundNumber shouldBe earliest
-        rootHash.value.rootHash should have length 64 // hex-encoded SHA-256
-      }
-
-      clue("Verify 404 for root hash of non-existent round") {
-        sv1ScanBackend.getRewardAccountingRootHash(earliest + 1) shouldBe None
-      }
-
-      val rootHashHex = sv1ScanBackend.getRewardAccountingRootHash(earliest).value.rootHash
-
-      clue("Verify batch lookup for root hash returns batch contents") {
-        val batch = sv1ScanBackend.getRewardAccountingBatch(earliest, rootHashHex)
-        batch shouldBe defined
-      }
-
-      clue("Verify 404 for batch lookup with non-existent hash") {
-        sv1ScanBackend.getRewardAccountingBatch(earliest, "0" * 64) shouldBe None
-      }
-    }
-  }
-}
