+++ test-full-class-names-sim-time.log
@@ -1,7 +1,6 @@
 org.lfdecentralizedtrust.splice.integration.tests.DisabledWalletTimeBasedIntegrationTest
 org.lfdecentralizedtrust.splice.integration.tests.ExternallySignedTxsTimeBasedIntegrationTest
 org.lfdecentralizedtrust.splice.integration.tests.FollowAmuletConversionRateFeedTimeBasedIntegrationTest
-org.lfdecentralizedtrust.splice.integration.tests.ScanAppRewardsComputationTimeBasedIntegrationTest
 org.lfdecentralizedtrust.splice.integration.tests.ScanWithGradualStartsTimeBasedIntegrationTest
 org.lfdecentralizedtrust.splice.integration.tests.SvExpiredRewardsCollectionTimeBasedIntegrationTest
 org.lfdecentralizedtrust.splice.integration.tests.SvTimeBasedAmuletPriceIntegrationTest
